pyproject.toml
[project]
name = "legacy-architect"
version = "0.1.0"
description = "Gemini 3-powered safe refactor agent for legacy code"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "google-genai>=1.0.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["legacy_architect*"]
